Deprecate `TitleLabel`, `SubtitleLabel`, `ManeuverContainerView` and `LineView`. Extract style related components to separate files to improve readability and minimize merge conflicts.
