+++ Sources/MapboxNavigation/BannerView.swift
@@ -1,0 +1,24 @@
+import UIKit
+
+// :nodoc:
+@objc(MBBannerContainerView)
+open class BannerContainerView: UIView {
+    
+}
+
+// :nodoc:
+@objc(MBTopBannerView)
+open class TopBannerView: UIView {
+    
+}
+
+// :nodoc:
+@objc(MBBottomBannerView)
+open class BottomBannerView: UIView {
+    
+}
+
+// :nodoc:
+open class BottomPaddingView: BottomBannerView {
+    
+}
