+++ Sources/MapboxNavigation/SeparatorView.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+// :nodoc:
+@objc(MBSeparatorView)
+public class SeparatorView: UIView {
+    
+}
