+++ Sources/MapboxNavigation/StylableView.swift
@@ -1,0 +1,28 @@
+import UIKit
+
+// :nodoc:
+@objc(MBStylableView)
+open class StylableView: UIView {
+    
+    @objc dynamic public var borderWidth: CGFloat = 0.0 {
+        didSet {
+            layer.borderWidth = borderWidth
+        }
+    }
+    
+    @objc dynamic public var cornerRadius: CGFloat = 0.0 {
+        didSet {
+            layer.cornerRadius = cornerRadius
+        }
+    }
+    
+    @objc dynamic public var borderColor: UIColor? {
+        get {
+            guard let color = layer.borderColor else { return nil }
+            return UIColor(cgColor: color)
+        }
+        set {
+            layer.borderColor = newValue?.cgColor
+        }
+    }
+}
