+++ Sources/MapboxNavigation/SubtitleLabel.swift
@@ -2,7 +2,7 @@ import UIKit
 
 // :nodoc:
 @available(*, deprecated, message: "This class is no longer used.")
-@objc(MBHighlightedButton)
-public class HighlightedButton: Button {
+@objc(MBSubtitleLabel)
+open class SubtitleLabel: StylableLabel {
 
 }
